backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
aiosqlite
databases
passlib
bcrypt
pyjwt
python-multipart